+++ src/main/java/ru/yandex/money/gradle/plugins/grafana/dashboard/GrafanaDashboardPlugin.java
@@ -1,23 +1,36 @@
 package ru.yandex.money.gradle.plugins.grafana.dashboard;
 
+import kotlin.KotlinVersion;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.DependencySet;
+import org.gradle.api.file.FileTree;
 import org.gradle.api.internal.artifacts.dependencies.DefaultExternalModuleDependency;
 import org.gradle.api.plugins.JavaBasePlugin;
 import org.gradle.api.plugins.JavaPluginConvention;
+import org.gradle.api.tasks.Copy;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.SourceSetContainer;
-import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
+import java.nio.file.Paths;
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Grafana dashboards plugin
  */
 @SuppressWarnings("ClassWithoutConstructor")
 public class GrafanaDashboardPlugin implements Plugin<Project> {
+    /**
+     * Directory, where dashboards from libraries will be saved
+     */
+    static final String DASHBOARDS_FROM_ARTIFACT_DIR = "grafana";
+    private static final String GRAFANA_ARTIFACT_SOURCE_SET_NAME = "grafanaFromArtifact";
+    private static final String GRAFANA_DIR_SOURCE_SET_NAME = "grafanaFromDir";
+    private static final String GRAFANA_DASHBOARDS_CONFIGURATION_NAME = "grafanaDashboards";
+    private static final String UPLOAD_TASK_NAME = "uploadGrafanaDashboards";
 
     /**
      * Actions when applying GrafanaDashboardPlugin to a project
@@ -26,11 +39,20 @@ public class GrafanaDashboardPlugin implements Plugin<Project> {
      */
     @Override
     public void apply(Project target) {
-        GrafanaDashboardExtension grafanaDashboardExtension = getGrafanaDashboardExtensionWithDefaults(target);
         target.getPluginManager().apply(JavaBasePlugin.class);
-        Configuration grafanaConfiguration = configureSourceSets(target, grafanaDashboardExtension);
+        GrafanaDashboardExtension grafanaDashboardExtension = getGrafanaDashboardExtensionWithDefaults(target);
+
+        Configuration grafanaFromDirConfiguration = configureDirSourceSets(target, grafanaDashboardExtension);
+        Configuration grafanaFromArtifactConfiguration = configureArtifactSourceSets(target);
+
+        Configuration grafanaDashboardsConfiguration = target.getConfigurations()
+                .maybeCreate(GRAFANA_DASHBOARDS_CONFIGURATION_NAME + "Compile");
+
         target.afterEvaluate(project -> {
-            createUploadGrafanaDashboardTask(project, grafanaConfiguration, grafanaDashboardExtension);
+            createUploadGrafanaDashboardTask(project,
+                    grafanaFromDirConfiguration, grafanaFromArtifactConfiguration, grafanaDashboardExtension);
+
+            createExtractGrafanaDashboardsTask(grafanaDashboardsConfiguration, target);
         });
     }
 
@@ -41,30 +63,60 @@ public class GrafanaDashboardPlugin implements Plugin<Project> {
         return grafanaDashboardExtension;
     }
 
-    private Configuration configureSourceSets(Project project, GrafanaDashboardExtension grafanaDashboardExtension) {
-
+    private Configuration configureDirSourceSets(Project project,
+                                                 GrafanaDashboardExtension grafanaDashboardExtension) {
         SourceSetContainer sourceSets = project.getConvention().getPlugin(JavaPluginConvention.class).getSourceSets();
-        SourceSet grafanaSourceset = sourceSets.create("grafana");
+        SourceSet grafanaSourceset = sourceSets.create(GRAFANA_DIR_SOURCE_SET_NAME);
         grafanaSourceset.getJava().srcDir(new File(grafanaDashboardExtension.dir));
 
-        Configuration grafanaCompileConfiguration = project.getConfigurations().getByName("grafanaCompile");
-        DependencySet grafanaCompileDependencies = grafanaCompileConfiguration.getDependencies();
-        grafanaCompileDependencies.add(new DefaultExternalModuleDependency(
-                "org.jetbrains.kotlin", "kotlin-stdlib", "1.2.61"));
-        grafanaCompileDependencies.add(new DefaultExternalModuleDependency(
-                "org.jetbrains.kotlin", "kotlin-reflect", "1.2.61"));
-        return grafanaCompileConfiguration;
+        Configuration grafanaDirCompile = project.getConfigurations()
+                .maybeCreate(GRAFANA_DIR_SOURCE_SET_NAME + "Compile");
+
+        return addKotlinDependencies(grafanaDirCompile);
     }
 
-    private static UploadGrafanaDashboardsTask createUploadGrafanaDashboardTask(
-            Project target, Configuration grafanaConfiguration, GrafanaDashboardExtension grafanaDashboardExtension) {
+    private Configuration configureArtifactSourceSets(Project project) {
+        SourceSetContainer sourceSets = project.getConvention().getPlugin(JavaPluginConvention.class).getSourceSets();
+        SourceSet grafanaSourceset = sourceSets.create(GRAFANA_ARTIFACT_SOURCE_SET_NAME);
+        grafanaSourceset.getJava()
+                .srcDir(Paths.get(project.getBuildDir().toString(), GRAFANA_ARTIFACT_SOURCE_SET_NAME).toString());
+
+        Configuration grafanaDirCompile = project.getConfigurations()
+                .maybeCreate(GRAFANA_ARTIFACT_SOURCE_SET_NAME + "Compile");
+
+        return addKotlinDependencies(grafanaDirCompile);
+    }
+
+    private Configuration addKotlinDependencies(Configuration grafanaConfiguration) {
+        DependencySet grafanaCompileDependencies = grafanaConfiguration.getDependencies();
+        grafanaCompileDependencies.add(new DefaultExternalModuleDependency(
+                "org.jetbrains.kotlin", "kotlin-stdlib", KotlinVersion.CURRENT.toString()));
+        grafanaCompileDependencies.add(new DefaultExternalModuleDependency(
+                "org.jetbrains.kotlin", "kotlin-reflect", KotlinVersion.CURRENT.toString()));
+        return grafanaConfiguration;
+    }
+
+    private void createExtractGrafanaDashboardsTask(Configuration grafanaDashboardsConfiguration, Project project) {
+        Copy task = project.getTasks().create("extractGrafanaDashboards", Copy.class);
+        List<FileTree> files = grafanaDashboardsConfiguration.getFiles().stream()
+                .map(project::zipTree)
+                .collect(Collectors.toList());
+        task.from(files);
+
+        task.into(Paths.get(project.getBuildDir().toString(), DASHBOARDS_FROM_ARTIFACT_DIR).toString());
+        project.getTasks().getByName(UPLOAD_TASK_NAME).dependsOn(task);
+    }
+
+    private static void createUploadGrafanaDashboardTask(
+            Project target, Configuration grafanaFromDirConfiguration,
+            Configuration grafanaFromArtifactConfiguration, GrafanaDashboardExtension grafanaDashboardExtension) {
         UploadGrafanaDashboardsTask task = target.getTasks()
-                .create("uploadGrafanaDashboards", UploadGrafanaDashboardsTask.class);
+                .create(UPLOAD_TASK_NAME, UploadGrafanaDashboardsTask.class);
         task.setGroup("other");
         task.setDescription("Upload Grafana dashboards");
-        task.setGrafanaConfiguration(grafanaConfiguration);
+        task.setGrafanaFromDirConfiguration(grafanaFromDirConfiguration);
+        task.setGrafanaFromArtifactConfiguration(grafanaFromArtifactConfiguration);
         task.setGrafanaDashboardExtension(grafanaDashboardExtension);
-        return task;
     }
 
 }