+++ src/main/java/ru/yandex/money/gradle/plugins/grafana/dashboard/impl/GrafanaDashboardUploader.java
@@ -7,10 +7,11 @@ import org.gradle.api.logging.Logging;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * Grafana dashboard uploader
@@ -42,9 +43,18 @@ public class GrafanaDashboardUploader {
             log.lifecycle("Grafana directory not found: dir={}", targetDir);
             return;
         }
-        File[] files = targetDir.listFiles(pathname -> dashboardContentCreators.stream()
-                .anyMatch(creator -> creator.isSupported(pathname)));
-        List<File> dashboards = files == null ? Collections.emptyList() : Arrays.asList(files);
+        List<File> dashboards;
+        try {
+            dashboards = Files.walk(targetDir.toPath())
+                    .filter(Files::isRegularFile)
+                    .map(Path::toFile)
+                    .filter(pathname -> dashboardContentCreators.stream()
+                            .anyMatch(creator -> creator.isSupported(pathname)))
+                    .collect(Collectors.toList());
+        } catch (IOException e) {
+            throw new RuntimeException("Can't get dashboard files", e);
+        }
+
         if (dashboards.isEmpty()) {
             log.lifecycle("No grafana dashboards: dir={}", targetDir.getAbsolutePath());
             return;