+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=MAJOR
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+1. Added grafanaDashboardsCompile sourceSet, where you can specify libraries which contains additional 
+dashboards.
+2. SourceSet was separated into grafanaFromArtifact and grafanaFromDir for the purpoce to announce different versions of dsl.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [2.2.2]() (22-05-2019)